chore(fe): combine carousel components
